Add runRight action
The action framework is getting a little more fleshed out. Now, key inputs set
the action of the Entity, which then informs the sprite rendering function.
These functions decide for themselves how to interpret and update the Entity's
frame, and can even change the action of the entity, if it decides that its
action is no longer applicable.